test(EditableView.test.tsx): remove unnecessary test case for delete functionality
The test case for the delete functionality has been removed as it was unnecessary and redundant. The test was checking for the presence of a button with the text "Delete" and then clicking on it to call a deleteContent mutation, which was not required for the current scope of testing.