chore(package.json): update devDependencies versions for testing libraries and vitest
The package name is updated to "@dictybase/google-analytics" to maintain consistency with other packages in the project. DevDependencies versions for testing libraries and vitest are updated to the latest versions for compatibility and potential improvements. The fp-ts dependency version is updated to "^2.16.1" to ensure compatibility with the latest features and bug fixes.