feat(contentPage.spec.ts): refactor contentPage.spec.ts to use functional programming with fp-ts library for better code readability and maintainability
The changes introduce the fp-ts library for functional programming in the contentPage.spec.ts file. The code is refactored to use functional programming concepts like piping, mapping, and flattening arrays to improve code readability and maintainability. This change aims to make the code more concise and easier to understand for developers familiar with functional programming paradigms.