chore(dicty-frontpage): add lint and lint:out scripts for linting with ESLint
The lint script runs ESLint on the src directory with .ts and .tsx file extensions, using the --cache option for improved performance. The lint:out script runs ESLint on the src directory with the same file extensions but outputs the results in JSON format to a specific file location. These scripts are added to enhance code quality by enforcing consistent coding standards and identifying potential issues early in the development process.