refactor(ci.yml): streamline deployment jobs into a single job for multiple apps
This change consolidates the deployment of the frontpage, stockcenter,
and publication applications into a single CI job. This reduces
redundancy and simplifies the CI configuration by using a single job to
deploy multiple related applications. The use of a combined docker image
and dockerfile parameter for all three applications further centralizes
the deployment process, making it more efficient and easier to manage.