feat(Dockerfile.frontpage): update base images for Node and Golang
The base image for Node.js is updated to `node:lts-alpine3.20` from `node:lts-alpine3.18` to leverage the latest LTS version, ensuring better security and performance. Similarly, the Golang image is updated to `golang:1.22.6-alpine3.20` from `golang:1.21.7-alpine3.19` to utilize the latest features and improvements in the Golang ecosystem. This also aligns with the updated Alpine version, maintaining consistency across the Docker environment.