feat(docker): update base images in Dockerfiles to newer versions
Updating the base images in Dockerfiles for both the publication and
stock-center services to use newer versions of node and golang. This
change ensures that the services are running on the latest supported and
more secure versions of these platforms. The update includes moving from
node:lts-alpine3.18 to node:lts-alpine3.20 and from older versions of
golang-alpine to golang:1.22.6-alpine3.20, which may include performance
improvements and security patches.