feat(ui-frontpage): add support for rendering dynamic content in AuthorizedDictyNews component
The changes include adding support for rendering dynamic content in the AuthorizedDictyNews component by updating the mockContent object with a new text value and modifying the test to expect the new text to be rendered on the screen. This enhancement allows for more flexible and customizable content to be displayed in the component.