feat(ui-dsc): add usePhenotypeValidation hook for form validation in the catalog component
The usePhenotypeValidation hook is added to provide form validation for the catalog component. It utilizes the react-hook-form library along with Yup schema validation to ensure that the phenotype and reference publication fields are required. Additionally, it includes a test to check if the publication exists by making a query to the GraphQL schema. This hook enhances the user experience by validating form inputs before submission.