feat(Dockerfile.genomepage): add Dockerfile for genome-page service to build and run the application in a containerized environment
The Dockerfile.genomepage has been added to the project to facilitate the containerization of the genome-page service. This Dockerfile defines the necessary steps to build and run the application within a Docker container. It includes setting up the base image, installing dependencies, building the project, and configuring the runtime environment for the application to run successfully in a containerized environment.