build(genome-page): update build script to include export command for static site generation
The build script in the package.json file for the genome-page app has been updated to include the "next export -o ./dist" command. This change enables the generation of a static version of the site in the ./dist directory, allowing for easier deployment and hosting of the application as a static site.