feat(ui-common): add SectionSelect component for selecting different sections in forms
feat(ui-common): export SectionSelect component for use in other parts of the application
The SectionSelect component is added to the ui-common package to allow users to select different sections in forms. This component provides a dropdown menu with predefined section options. Additionally, the SectionSelect component is exported for use in other parts of the application, enhancing reusability and modularity.