feat(useValidateCreateContent): add custom hook useValidateCreateContent for form validation
The useValidateCreateContent custom hook is added to handle form validation for creating content. It uses the yup library for defining validation schema rules for the form fields. Additionally, it integrates with react-hook-form for form management and validation. The hook also includes a test function to check if the content already exists based on the provided section, name, and subname values. This custom hook enhances code reusability and maintainability by encapsulating the validation logic in a separate module.