feat(create.tsx): add a new CreateContentView component for creating content with an editor and form
The changes introduce a new file, create.tsx, which contains the CreateContentView component. This component includes an Editor component from "@dictybase/editor" for editing content and a CreateContentForm component for additional form functionality. The component is designed to be used for creating content and is set to be private with roles restricted to "content-admin".