fix(PlasmidDetailsCard.tsx): improve type safety
Previously, the types of values being passed as the cartData property
were being asserted using the `as` keyword, which are not guaranteed at
runtime. The summary variable of the plasmid is now given a default
value of an empty string if the value is null or undefined to ensure
that is the correct type.